refactor: moved key stuff to own mixin
