+++ src/dataLoadingMixin.js
@@ -1,7 +1,4 @@
 import { checkFormatColumnData, checkFormatInternal } from './utils/checkFormat.js'
-import { generateKeyColumn, validateKeyColumn } from './utils/key.js'
-
-import getDataLength from './utils/getDataLength.js'
 import convertRowToColumnData from './utils/convertRowToColumnData.js'
 import parseGeoJSON from './utils/parseGeoJSON.js'
 
@@ -39,33 +36,6 @@ const methods = {
 
     if (options.validate === true) {
       this.validateAllColumns()
-    }
-  },
-
-  _setupKeyColumn () {
-    const length = getDataLength(this._data)
-
-    if ('$key' in this._data) {
-      validateKeyColumn(this._data.$key, length)
-      this._syncKeyToRowNumber()
-    } else {
-      const keyColumn = generateKeyColumn(length)
-      this._setKeyColumn(keyColumn)
-    }
-  },
-
-  _setKeyColumn (keyColumn) {
-    this._data.$key = keyColumn
-
-    this._syncKeyToRowNumber()
-  },
-
-  _syncKeyToRowNumber () {
-    const length = getDataLength(this._data)
-
-    for (let i = 0; i < length; i++) {
-      const key = this._data.$key[i]
-      this._keyToRowNumber[key] = i
     }
   }
 }
